backend/requirements.txt
Django==2.2.19
pytz==2023.3.post1
sqlparse==0.4.4
djangorestframework==3.12.4